+++ contracts/mocks/sale/SaleMock.sol
@@ -30,7 +30,7 @@ contract SaleMock is Sale {
         bytes32 sku,
         IERC20 token
     ) external view returns (uint256 price) {
-        price = _skuTokenPrices.getPrice(sku, token);
+        price = _getPrice(sku, token);
     }
 
     function callUnderscorePurchaseFor(