+++ contracts/mocks/sale/SimpleSaleMock.sol
@@ -26,7 +26,7 @@ contract SimpleSaleMock is SimpleSale {
         bytes32 sku,
         IERC20 token
     ) external view returns (uint256 price) {
-        price = _skuTokenPrices.getPrice(sku, token);
+        price = _getPrice(sku, token);
     }
 
     function callUnderscoreValidatePurchase(