+++ contracts/sale/pricing/SkuTokenPrice.sol
@@ -7,7 +7,8 @@ import "@animoca/ethereum-contracts-core_library/contracts/algo/EnumSet.sol";
 
 /**
  * @title SkuTokenPrice
- * A library contract that manages the token prices for a set of product SKUs.
+ * A contract module that adds support for managing the token prices for a set
+ * of product SKUs.
  *
  * Sku Token Price managers have the following properites:
  *
@@ -15,41 +16,27 @@ import "@animoca/ethereum-contracts-core_library/contracts/algo/EnumSet.sol";
  * - There are no guarantees on the ordering of elements in the list of SKUs and
  *  tokens.
  * - Set and get the price for a given SKU and token.
- *
- * ```
- * contract Example {
- *     // Add the library methods
- *     using SkuTokenPrice for SkuTokenPrice.Manager;
- *
- *     // Declare a sku token price manager state variable
- *     SkuTokenPrice.Manager private mySkuTokenPriceManager;
- * }
- * ```
  */
-library SkuTokenPrice {
+contract SkuTokenPrice {
 
     using EnumSet for EnumSet.Set;
 
-    struct Manager {
-        // mapping of SKUs to their mapping of token prices
-        mapping(bytes32 => mapping(IERC20 => uint256)) skuTokenPrices;
+    // mapping of SKUs to their mapping of token prices
+    mapping(bytes32 => mapping(IERC20 => uint256)) private _skuTokenPrices;
 
-        // list of supported SKUs
-        EnumSet.Set skus;
+    // list of supported SKUs
+    EnumSet.Set private _skus;
 
-        // list of supported tokens
-        EnumSet.Set tokens;
-    }
+    // list of supported tokens
+    EnumSet.Set private _tokens;
 
     /**
      * Adds a batch of SKUs to the list of supported product SKUs.
-     * @param manager The storage instance of the SkuTokenPrice manager.
      * @param skus The list of additional SKUs to support.
      * @return added A list of flags, corresponding to the input list of SKUs,
      *  indicating whether or not each SKU list element has been added.
      */
-    function addSkus(
-        Manager storage manager,
+    function _addSkus(
         bytes32[] memory skus
     )
         internal
@@ -61,19 +48,17 @@ library SkuTokenPrice {
 
         for (uint256 index = 0; index < numSkus; ++index) {
             bytes32 sku = skus[index];
-            added[index] = manager.skus.add(sku);
+            added[index] = _skus.add(sku);
         }
     }
 
     /**
      * Removes a batch of SKUs from the list of supported product SKUs.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      * @param skus The list of SKUs to remove.
      * @return removed A list of flags, corresponding to the input list of SKUs,
      *  indicating whether or not each SKU list element has been removed.
      */
-    function removeSkus(
-        Manager storage manager,
+    function _removeSkus(
         bytes32[] memory skus
     )
         internal
@@ -86,12 +71,12 @@ library SkuTokenPrice {
         for (uint256 skuIndex = 0; skuIndex < numSkus; ++skuIndex) {
             bytes32 sku = skus[skuIndex];
 
-            if (manager.skus.remove(sku)) {
-                uint256 numTokens = manager.tokens.length();
+            if (_skus.remove(sku)) {
+                uint256 numTokens = _tokens.length();
 
                 for (uint256 tokenIndex = 0; tokenIndex != numTokens; ++tokenIndex) {
-                    IERC20 token = IERC20(address(uint256(manager.tokens.at(tokenIndex))));
-                    delete manager.skuTokenPrices[sku][token];
+                    IERC20 token = IERC20(address(uint256(_tokens.at(tokenIndex))));
+                    delete _skuTokenPrices[sku][token];
                 }
 
                 removed[skuIndex] = true;
@@ -103,48 +88,41 @@ library SkuTokenPrice {
 
     /**
      * Validates whether or not the specified SKU is supported.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      * @param sku The SKU to validate.
      * @return exists True if the specified SKU is supported, false otherwise.
      */
-    function hasSku(
-        Manager storage manager,
+    function _hasSku(
         bytes32 sku
     )
         internal view
         returns (bool exists)
     {
-        exists = manager.skus.contains(sku);
+        exists = _skus.contains(sku);
     }
 
     /**
      * Retrieves the entire list of supported SKUs.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      */
-    function getSkus(
-        Manager storage manager
-    )
+    function _getSkus()
         internal view
         returns (bytes32[] memory skus)
     {
-        uint256 numSkus = manager.skus.length();
+        uint256 numSkus = _skus.length();
 
         skus = new bytes32[](numSkus);
 
         for (uint256 index = 0; index != numSkus; ++index) {
-            skus[index] = manager.skus.at(index);
+            skus[index] = _skus.at(index);
         }
     }
 
     /**
      * Adds a batch of IERC20 tokens to the list of supported tokens.
-     * @param manager The storage instance of the SkuTokenPrice manager.
      * @param tokens The list of additional IERC20 tokens to support.
      * @return added A list of flags, corresponding to the input list of tokens,
      *  indicating whether or not each token list element has been added.
      */
-    function addTokens(
-        Manager storage manager,
+    function _addTokens(
         IERC20[] memory tokens
     )
         internal
@@ -156,20 +134,18 @@ library SkuTokenPrice {
 
         for (uint256 index = 0; index < numTokens; ++index) {
             IERC20 token = tokens[index];
-            added[index] = manager.tokens.add(bytes32(uint256(address(token))));
+            added[index] = _tokens.add(bytes32(uint256(address(token))));
         }
     }
 
     /**
      * Removes a batch of IERC20 tokens from the list of supported tokens.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      * @param tokens The list of IERC20 tokens to remove.
      * @return removed A list of flags, corresponding to the input list of
      *  IERC20 tokens, indicating whether or not each token list element has been
      *  removed.
      */
-    function removeTokens(
-        Manager storage manager,
+    function _removeTokens(
         IERC20[] memory tokens
     )
         internal
@@ -182,12 +158,12 @@ library SkuTokenPrice {
         for (uint256 tokenIndex = 0; tokenIndex < numTokens; ++tokenIndex) {
             IERC20 token = tokens[tokenIndex];
 
-            if (manager.tokens.remove(bytes32(uint256(address(token))))) {
-                uint256 numSkus = manager.skus.length();
+            if (_tokens.remove(bytes32(uint256(address(token))))) {
+                uint256 numSkus = _skus.length();
 
                 for (uint256 skuIndex = 0; skuIndex != numSkus; ++skuIndex) {
-                    bytes32 sku = manager.skus.at(skuIndex);
-                    delete manager.skuTokenPrices[sku][token];
+                    bytes32 sku = _skus.at(skuIndex);
+                    delete _skuTokenPrices[sku][token];
                 }
 
                 removed[tokenIndex] = true;
@@ -199,36 +175,31 @@ library SkuTokenPrice {
 
     /**
      * Validates whether or not the specified ERC20 token is supported.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      * @param token The ERC20 token to validate.
      * @return exists True if the specified token is supported, false otherwise.
      */
-    function hasToken(
-        Manager storage manager,
+    function _hasToken(
         IERC20 token
     )
         internal view
         returns (bool exists)
     {
-        exists = manager.tokens.contains(bytes32(uint256(address(token))));
+        exists = _tokens.contains(bytes32(uint256(address(token))));
     }
 
     /**
      * Retrieves the entire list of supported ERC20 tokens.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      */
-    function getTokens(
-        Manager storage manager
-    )
+    function _getTokens()
         internal view
         returns (IERC20[] memory tokens)
     {
-        uint256 numTokens = manager.tokens.length();
+        uint256 numTokens = _tokens.length();
 
         tokens = new IERC20[](numTokens);
 
         for (uint256 index = 0; index != numTokens; ++index) {
-            tokens[index] = IERC20(address(uint256(manager.tokens.at(index))));
+            tokens[index] = IERC20(address(uint256(_tokens.at(index))));
         }
     }
 
@@ -236,14 +207,12 @@ library SkuTokenPrice {
      * Retrieves the price for the specified supported SKU and token.
      * @dev Reverts if the specified SKU does not exist.
      * @dev Reverts is the specified ERC20 token is unsupported.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      * @param sku The SKU item whose token price will be retrieved.
      * @param token The ERC20 token whose SKU price will be retrieved.
      * @return price The retrieved price for the specified supported SKU and
      *  token.
      */
-    function getPrice(
-        Manager storage manager,
+    function _getPrice(
         bytes32 sku,
         IERC20 token
     )
@@ -251,14 +220,14 @@ library SkuTokenPrice {
         returns (uint256 price)
     {
         require(
-            manager.skus.contains(sku),
+            _skus.contains(sku),
             "SkuTokenPrice: non-existent sku");
 
         require(
-            manager.tokens.contains(bytes32(uint256(address(token)))),
+            _tokens.contains(bytes32(uint256(address(token)))),
             "SkuTokenPrice: unsupported token");
 
-        price = manager.skuTokenPrices[sku][token];
+        price = _skuTokenPrices[sku][token];
     }
 
     /**
@@ -266,14 +235,12 @@ library SkuTokenPrice {
      * @dev Reverts if the specified SKU does not exist.
      * @dev Reverts if the token/price list lengths are not aligned.
      * @dev Reverts if any of the specified ERC20 tokens are unsupported.
-     * @param manager The storage instance of the SkuTokenPriceManager.
      * @param sku The SKU item whose token price will be set.
      * @param tokens The list of ERC20 tokens whose SKU price will be set.
      * @param prices The list of prices to set with.
      * @return prevPrices The list of previous SKU token prices.
      */
-    function setPrices(
-        Manager storage manager,
+    function _setPrices(
         bytes32 sku,
         IERC20[] memory tokens,
         uint256[] memory prices
@@ -282,7 +249,7 @@ library SkuTokenPrice {
         returns (uint256[] memory prevPrices)
     {
         require(
-            manager.skus.contains(sku),
+            _skus.contains(sku),
             "SkuTokenPrice: non-existent sku");
 
         require(
@@ -297,12 +264,12 @@ library SkuTokenPrice {
             IERC20 token = tokens[index];
 
             require(
-                manager.tokens.contains(bytes32(uint256(address(token)))),
+                _tokens.contains(bytes32(uint256(address(token)))),
                 "SkuTokenPrice: unsupported token");
 
-            prevPrices[index] = manager.skuTokenPrices[sku][token];
+            prevPrices[index] = _skuTokenPrices[sku][token];
 
-            manager.skuTokenPrices[sku][token] = prices[index];
+            _skuTokenPrices[sku][token] = prices[index];
         }
     }